Working on map code to display pins with desired path from starting point to destination. Not final code and still need to clean it up
