+++ SwiftUI-AR-School-App/SwiftUI_AR_School_AppApp.swift
@@ -8,6 +8,7 @@
 import SwiftUI
 
 @main
+@available(iOS 17.0, *)
 struct SwiftUI_AR_School_AppApp: App {
     var body: some Scene {
         WindowGroup {
